refector: refector create project's main function
