+++ create/test/utils/getWantedFeature.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect, vi } from "vitest";
 
-import { getWantedFeature } from "@/utils/prompts/getWantedFeature";
+import { getWantedFeature } from "@/utils/propmpt/getWantedFeature";
 
 describe("getWantedFeature()", () => {
   test("Return an wantedFeatures object with empty array if isFullFeature is false.", async () => {
