+++ create/utils/prompt/getIsDescription.ts
